# web/fixtures/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Standards360",
    "summary": "Maps a procurement specification to the applicable Indian Standards, their allied standards, current versions and certification obligations.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/analyze": {
      "post": {
        "summary": "Analyze",
        "description": "Analyse pasted text, a specification, or a bare product name.",
        "operationId": "analyze_v1_analyze_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalyzeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalyzeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/analyze/upload": {
      "post": {
        "summary": "Analyze Upload",
        "description": "Analyse an uploaded tender: PDF, DOCX, text file, or a screenshot.",
        "operationId": "analyze_upload_v1_analyze_upload_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_analyze_upload_v1_analyze_upload_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalyzeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/document": {
      "post": {
        "summary": "Document",
        "description": "Generate the PDF for pasted text: a standalone report, or a bare annexure.",
        "operationId": "document_v1_document_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DocumentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/document/upload": {
      "post": {
        "summary": "Document Upload",
        "description": "Analyse an uploaded tender and return it with the annexure appended.\n\nOutput mode 1 from the manual: the officer's own document comes back complete, which is how a\ntender actually carries this information.",
        "operationId": "document_upload_v1_document_upload_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_document_upload_v1_document_upload_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/standard/{is_number}": {
      "get": {
        "summary": "Standard",
        "description": "Everything the catalogue holds about one standard, for a details panel in the interface.",
        "operationId": "standard_v1_standard__is_number__get",
        "parameters": [
          {
            "name": "is_number",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Is Number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AnalyzeRequest": {
        "properties": {
          "text": {
            "type": "string",
            "title": "Text",
            "description": "Tender text, a specification, or a product name"
          },
          "lang": {
            "type": "string",
            "title": "Lang",
            "description": "Language of the text; 'auto' for detection",
            "default": "auto"
          },
          "persona": {
            "type": "string",
            "title": "Persona",
            "description": "procurement | manufacturer",
            "default": "procurement"
          },
          "output_mode": {
            "type": "string",
            "title": "Output Mode",
            "description": "annexure | report",
            "default": "annexure"
          },
          "state": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "State",
            "description": "Used to list the nearest testing laboratories"
          },
          "answers": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "string"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Answers",
            "description": "Replies to the questions of a previous call"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "text"
        ],
        "title": "AnalyzeRequest"
      },
      "AnalyzeResponse": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status",
            "description": "complete | need_more_info | no_match"
          },
          "query": {
            "type": "string",
            "title": "Query",
            "default": ""
          },
          "requirement": {
            "additionalProperties": true,
            "type": "object",
            "title": "Requirement"
          },
          "questions": {
            "items": {
              "additionalProperties": {
                "type": "string"
              },
              "type": "object"
            },
            "type": "array",
            "title": "Questions"
          },
          "confidence": {
            "type": "number",
            "title": "Confidence",
            "default": 0.0
          },
          "band": {
            "type": "string",
            "title": "Band",
            "default": "low"
          },
          "confidence_drivers": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Confidence Drivers"
          },
          "primary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Primary"
          },
          "primary_title": {
            "type": "string",
            "title": "Primary Title",
            "default": ""
          },
          "primary_as_cited": {
            "type": "string",
            "title": "Primary As Cited",
            "default": ""
          },
          "amendments": {
            "items": {
              "additionalProperties": {
                "type": "string"
              },
              "type": "object"
            },
            "type": "array",
            "title": "Amendments"
          },
          "options": {
            "items": {
              "$ref": "#/components/schemas/OptionOut"
            },
            "type": "array",
            "title": "Options"
          },
          "allied": {
            "additionalProperties": {
              "items": {
                "additionalProperties": true,
                "type": "object"
              },
              "type": "array"
            },
            "type": "object",
            "title": "Allied"
          },
          "warnings": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Warnings"
          },
          "verdicts": {
            "items": {
              "$ref": "#/components/schemas/CitationVerdict"
            },
            "type": "array",
            "title": "Verdicts"
          },
          "certification": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CertificationAnswer"
              },
              {
                "type": "null"
              }
            ]
          },
          "evidence": {
            "items": {
              "$ref": "#/components/schemas/EvidenceOut"
            },
            "type": "array",
            "title": "Evidence"
          },
          "explanation": {
            "type": "string",
            "title": "Explanation",
            "default": ""
          },
          "removed_by_guard": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Removed By Guard"
          },
          "pdf_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pdf Url",
            "description": "Filled once D3 generates the document"
          },
          "seconds": {
            "type": "number",
            "title": "Seconds",
            "default": 0.0
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "status"
        ],
        "title": "AnalyzeResponse",
        "description": "The only contract the front end and a procurement portal ever see (manual \u00a79)."
      },
      "Body_analyze_upload_v1_analyze_upload_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          },
          "persona": {
            "type": "string",
            "title": "Persona",
            "default": "procurement"
          },
          "state": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "State"
          },
          "lang": {
            "type": "string",
            "title": "Lang",
            "default": "auto"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_analyze_upload_v1_analyze_upload_post"
      },
      "Body_document_upload_v1_document_upload_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          },
          "persona": {
            "type": "string",
            "title": "Persona",
            "default": "procurement"
          },
          "option_id": {
            "type": "string",
            "title": "Option Id",
            "default": "B"
          },
          "mode": {
            "type": "string",
            "title": "Mode",
            "default": "annexure"
          },
          "reference": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reference"
          },
          "state": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "State"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_document_upload_v1_document_upload_post"
      },
...
      "DocumentRequest": {
        "properties": {
          "text": {
            "type": "string",
            "title": "Text",
            "description": "Tender text, a specification, or a product name"
          },
          "lang": {
            "type": "string",
            "title": "Lang",
            "description": "Language of the text; 'auto' for detection",
            "default": "auto"
          },
          "persona": {
            "type": "string",
            "title": "Persona",
            "description": "procurement | manufacturer",
            "default": "procurement"
          },
          "output_mode": {
            "type": "string",
            "title": "Output Mode",
            "description": "annexure | report",
            "default": "annexure"
          },
          "state": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "State",
            "description": "Used to list the nearest testing laboratories"
          },
          "answers": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "string"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Answers",
            "description": "Replies to the questions of a previous call"
          },
          "option_id": {
            "type": "string",
            "title": "Option Id",
            "description": "Citation depth: A, B or C",
            "default": "B"
          },
          "mode": {
            "type": "string",
            "title": "Mode",
            "description": "annexure | report",
            "default": "report"
          },
          "reference": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reference",
            "description": "Tender reference printed on the document"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "text"
        ],
        "title": "DocumentRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
    }
  }
}